Fix for trainers showing unlearnable abilities
Trainers are showing unlearnable abilities as learnable after learning
an ability or other action, such as receiving money, choosing a talent.

See Issue #2144 https://github.com/cmangos/issues/issues/2114

There are two issues:
1) spell prerequisites are not being communicated to client
2) for skill rank requirements, client ignores the requirement if skill
   not present in player.

Fix for (1) is in NPCHandler.cpp.  The spell chain prerequisites were
not being sent to the client. A check for !IsCastable(spellId) was
preventing the found prerequisites from being sent to the client as
all trainer (learn spell) spells are castable and the check always failed.
This check was removed.  Also, the learned spell was being calculated
locally ("triggerSpell" variable) instead of using the pre-calculated
value from TrainerSpell struct as done in TBC core.

Fix for (2) is more complex.  Apparently the client needs any skills
that need to be checked for rank requirements to be present in the
player data, iow sent to the client as if they were "learned" already.
The trick is that the skills are sent with zero rank and aren't shown
in the GUI to the user but are considered for prerequisite checks. This
required more intrusive changes to the skill system in Player.cpp.

As part of testing (2) found an issue with spells not getting unlearned
upon a profession being unlearned in Player::UpdateSkillTrainedSpells()
Just a headups in case this fix messes up something else.